Basic code to update react elements of the feed with GET requests
This code makes a GET request to the petfinder api & updates the state of react elements based on the results. There are still many things that need to get fixed with this code. The authentication code does not get updated upon invalidation (see authorizatonKey file for how to get authorization keys), there are no filters, and the pets do not get saved yet.  This code also need more error handeling & needs to be cleaned up a bit.