Updated the Feed files & got rid of the Feed folder
Converted the feed files to be rendered in feed.html. The scripts added to feed.html will need to be changed eventually for deployment but they work for no. I also added some CSS styling to  the react component, but did not add the media queries or on hover CSS.