Created the basic js for the pet page
I created the basic javascript to render a pet's information onto a page. The info being rendered is very simple and will be added to in the future. For now, I just based it off of the elements originally in the pet-page.html file. BEFORE LOADING THE pet-page.html FILE, THE setPetID FUNCTION SHOULD ALWAYS BE RAN! This sets what pet should be displayed on the page.