Fixed API request to get authroization tokens
The request in getToken() has been fixed. This function now returns a promise when the token has been set. When the function returns, any get request that failed due to the invalid token can be called again in a then statement.  pet-page.js  reactFeed.js have been updated to call this function properly on error of a get request.