Modified how pet-page loads info & added babble to package.json
Modified the pet-page code to use a URL with a id parameter to load a pet on to an html page & modified package.json to include a babble module that allows us to percompile the .js files that uses babble, so we dont have to include babble in the html files. To do this, all you need to do is move the .js files into the src folder, then run this command:

npx babel --watch src --out-dir . --presets react-app/prod

This will compile the code & add the file outside the src file.