Completed Color Validation
Color Validation was implemented, but the code doesn't alert the user of any invaild data they typed. This will be implemented later. The filters are getting passed into local storage. In order to get the filters to work correctly
account.html had to be modified slightly. It turns out the API only lets you look at one type of animal or all of them. The html was changed to list all animal types seperatly to follow this requirement.