Added a way to clear & load filters and a message display on update of filters
I added functions to clear the filters from local storage. This is called in a reset button. THIS BUTTON WILL PROBABLY NEED RESTYLING! I also added a way to load a users priviously saved filters into the form. The function gets called anytime the account.js file is loaded. Finally, I added a message that gets displayed after the filters get updated. This message also lists invalid data items to the user so they know the are not getting saved. Right now the message is displayed in an alert box, but this can easily be changed by editing line 167 of account.js